♻️ refactor: update project structure 📦 chore: add tauri to project 👽️ feat: requests directly in the app
This commit refactors the project structure to the project.
Which also includes new components, new pages, new contexts, and new providers.
Most importantly, this commit also adds the Tauri framework to the project, which will be used to create the desktop application.

This commit also includes a new component that is responsible for making requests directly in the application.
The component is responsible for making requests to the APIs that the application uses and then returning the data to the other components.
This way we can make requests directly in the application without the need for a backend server, which makes the requests faster.